# api/app/src/swagger/swagger_spec.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.0.1",
        "title": "MethodFitness"
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/fetchAppointments/{startDate}/{endDate}": {
            "get": {
                "x-name": "fetchAppointments",
                "description": "fetch Appointments",
                "operationId": "fetchAppointments",
                "parameters": [
                    {
                        "name": "startDate",
                        "in": "path",
                        "required": true,
                        "description": "the beginning of the span of time to retrieve appointments for",
                        "type": "string"
                    },
                    {
                        "name": "endDate",
                        "in": "path",
                        "required": true,
                        "description": "the ending of the span of time to retrieve appointments for",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/appointmentsResponse"
                        }
                    }
                }
            }
        },
        "/fetchAppointments/{startDate}/{endDate}/{trainerId}": {
            "get": {
                "x-name": "fetchAppointments",
                "description": "fetch Appointments",
                "operationId": "fetchAppointments",
                "parameters": [
                    {
                        "name": "startDate",
                        "in": "path",
                        "required": true,
                        "description": "the beginning of the span of time to retrieve appointments for",
                        "type": "string"
                    },
                    {
                        "name": "endDate",
                        "in": "path",
                        "required": true,
                        "description": "the ending of the span of time to retrieve appointments for",
                        "type": "string"
                    },
                    {
                        "name": "trainerId",
                        "in": "path",
                        "required": false,
                        "description": "the trainer id for whom to retrieve appointments",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/appointmentsResponse"
                        }
                    }
                }
            }
        },
        "/appointment/scheduleAppointment": {
            "post": {
                "x-name": "/appointment/scheduleAppointment",
                "description": "schedule Appointment",
                "operationId": "/appointment/scheduleAppointment",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scheduleAppointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/appointment/updateAppointment": {
            "post": {
                "x-name": "/appointment/updateAppointment",
                "description": "update Appointment",
                "operationId": "/appointment/updateAppointment",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scheduleAppointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
        "/appointment/cancelAppointment": {
            "post": {
                "x-name": "/appointment/cancelAppointment",
                "description": "cancel Appointment",
                "operationId": "/appointment/cancelAppointment",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cancelAppointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
...
        "/trainer/getTrainer/{id}": {
            "get": {
                "x-name": "trainer",
                "description": "retrieve single trainer by id",
                "operationId": "trainer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The trainers id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/standardSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/standardFailureResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "standardSuccessResponse": {
            "title": "standardResponse",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "payload": {
                    "type": "object"
                }
            }
        },
        "standardFailureResponse": {
            "title": "standardResponse",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "error": {
                    "type": "boolean"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "title": "errorMessage",
                        "properties": {
                            "field": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "appointmentsResponse": {
            "title": "appointmentsResponse",
            "properties": {
                "trainers": {
                    "type": "array",
                    "items": {
                        "title": "appointment",
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "appointmentType": {
                                "type": "string",
                                "x-input": "select"
                            },
                            "date": {
                                "type": "string",
                                "format": "date-time",
                                "x-input": "date-time"
                            },
                            "startTime": {
                                "type": "string",
                                "x-input": "select"
                            },
                            "endTime": {
                                "type": "string"
                            },
                            "trainer": {
                                "type": "string",
                                "format": "uuid",
                                "x-input": "select"
                            },
                            "clients": {
                                "type": "array",
                                "minItems": 1,
                                "items": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "uniqueItems": true,
                                "x-input": "multi-select"
                            },
                            "notes": {
                                "type": "string",
                                "x-input": "textArea"
                            }
                        },
                        "required": [
                            "appointmentType",
                            "date",
                            "startTime",
                            "endTime",
                            "trainer"
                        ]
                    }
                }
            }
        },
...
        "scheduleAppointment": {
            "title": "scheduleAppointment",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "appointmentId": {
                    "type": "string"
                },
                "appointmentType": {
                    "type": "string"
                },
                "date": {
                    "type": "string",
                    "format": "date-time"
                },
                "startTime": {
                    "type": "string"
                },
                "endTime": {
                    "type": "string"
                },
                "trainer": {
                    "type": "string"
                },
                "clients": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "uniqueItems": true
                },
                "notes": {
                    "type": "string"
                }
            },
            "required": [
                "appointmentType",
                "date",
                "startTime",
                "endTime",
                "trainer"
            ]
        },
        "cancelAppointment": {
            "title": "cancelAppointment",
            "properties": {
                "appointmentId": {
                    "type": "string",
                    "format": "uuid"
                },
                "entityName": {
                    "type": "string"
                }
            }
        },
...
    },
...
    "securityDefinitions": {},
...
}